moved progress bar to separate file
